backend/v1/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
flake8==6.1.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector-python==8.1.0
mysqlclient==2.0.3
protobuf==4.21.12
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
